standalone_ai_hub/requirements.txt
openai==1.35.3
python-dotenv==1.0.0
requests==2.31.0
pillow==10.3.0
flask==3.0.3
